* ⬆️ chore(package.json): upgrade dependencies Upgraded dependencies to their latest versions. Specifically, "@taquito/beacon-wallet", "@taquito/taquito", and "@taquito/utils" were upgraded to version "^16.1.2". "@testing-library/react" was upgraded to version "^14.0.0". "@types/node" was upgraded to version "^20.1.1". "typescript" was upgraded to version "^5.0.4".
* 🎨 style(App.tsx): change wording of tab labels
* 🐛 fix(App.tsx): change contract address to correct address
* 🚀 feat(App.tsx): add storage display to contract tab
The wording of the tab labels has been changed to better reflect the functionality of each tab. The contract address has been updated to the correct address. The contract tab now displays the current storage value of the NFT contract. This improves the user experience by providing more information about the current state of the contract.

* 🐛 fix(ConnectWallet.tsx): change setStorage type from number to any
* 🚀 chore(ConnectWallet.tsx): update BeaconWallet name to "Sample NFT dApp"
The setStorage function was previously set to accept a number, but it is now set to accept any type. This change was made to accommodate the new storage object that is being returned from the contract. The BeaconWallet name was also updated to better reflect the purpose of the application.

* 🐛 fix(Transfers.tsx): change amount variable to tokenId
* ✨ feat(Transfers.tsx): add support for token transfers
The amount variable has been changed to tokenId to better reflect the purpose of the variable. The component now supports token transfers by adding a new input field for the tokenId. The sendTransfer function has been updated to use the contract object to call the transfer method with the appropriate parameters. This allows the component to transfer tokens instead of just tez.

* ✨ feat(UpdateContract.tsx): add minting functionality
* 🐛 fix(UpdateContract.tsx): update setStorage to accept any type
The minting functionality has been added to the UpdateContract component. The mint function is called with the user's address and metadata. The metadata is an IPFS hash that is converted to bytes using the char2Bytes function from the Taquito library. The setStorage function has been updated to accept any type of data.

* 🚀 feat(UpdateContract.tsx): replace increment and decrement with mint and burn
* 🐛 fix(UpdateContract.tsx): update setStorage to set newStorage object
The increment and decrement functions have been replaced with mint and burn functions respectively. The mint function is used to add a new token to the contract, while the burn function is used to remove a token from the contract. The setStorage function has been updated to set the newStorage object instead of the newStorage.toNumber() value.